+++ src/singletons/Settings.hpp
@@ -320,6 +320,11 @@ public:
         true,
     };
 
+    IntSetting messageTruncationLimit = {
+        "/behaviour/messageTruncationLimit",
+        50,
+    };
+
     // Auto-completion
     BoolSetting onlyFetchChattersForSmallerStreamers = {
         "/behaviour/autocompletion/onlyFetchChattersForSmallerStreamers", true};